requirements.txt
PyQt5>=5.15
opencv-python
pillow
pytesseract
python-dotenv
numpy
requests
fuzzywuzzy
pynput
